requirements.txt
connected-components-3d==3.3.0
matplotlib
numpy==1.21.0
opencv-python==4.4.0.46
openpyxl
pandas
path
pathlib2==2.3.5
pathtools==0.1.2
PyYAML==5.3.1
scikit-image==0.18.2
scikit-learn
scipy
SimpleITK==1.2.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
torch==1.8.1
torch-model-archiver==0.1.0b20200409
torchio==0.18.39
torchserve==0.0.1b20200409
torchvision==0.9.1
tqdm
